+++ internal/delivery/http/router/auth_route.go
@@ -1,0 +1,17 @@
+package router
+
+import (
+	"github.com/afifudin23/absensi-king-royal-api/internal/delivery/http/handler"
+	"github.com/afifudin23/absensi-king-royal-api/internal/service"
+	"github.com/gin-gonic/gin"
+)
+
+func registerAuthRoutes(rg *gin.RouterGroup) {
+	authService := service.NewAuthService()
+	authHandler := handler.NewAuthHandler(authService)
+
+	auth := rg.Group("/auth")
+	auth.POST("/register", authHandler.Register)
+	auth.POST("/login", authHandler.Login)
+	auth.POST("/logout", authHandler.Logout)
+}